+++ frontend/src/components/specific/Search.vue
@@ -1,12 +1,11 @@
 <template>
-  <div class="search-container">
-    <input type="text" class="search-input" />
-    <button><Icons icon="search" /></button>
+  <div class="v-search">
+    <input type="text" class="v-search-input" placeholder="Sök" />
+    <button class="v-search-button"><Icons icon="search" /></button>
   </div>
 </template>
 
 <script setup lang="ts">
-import * as helper from '../../helper'
 import Icons from '../generic/Icons.vue'
 import '../../styles/components/search.scss'
 </script>
