+++ frontend/src/login_register/Login_Register.vue
@@ -1,5 +1,5 @@
 <template>
-	<Modal style="display: flex; justify-content: center; align-items: center">
+	<Modal v-if="!user.loggedIn" style="display: flex; justify-content: center; align-items: center">
 		<component :is="components[data?.type]"></component>
 	</Modal>
   </template>
@@ -8,6 +8,9 @@
 import login from './Login.vue'
 import register from './Register.vue'
 import Modal from '../components/generic/Modal.vue'
+import { useUserStore } from '../stores/user'
+
+const user = useUserStore()
 
 const props = defineProps({
   data: Object,
