+++ frontend/src/styles/components/listingitem.scss
@@ -1,25 +1,19 @@
-.item {
-	&-wrapper {
-		width: min(250px);
-		height: min(300px);
-		margin: 16px;
-		border-radius: 16px;
-		box-shadow: 0px 0px 5px 1px black;
-		transition: 0.1s;
-		overflow: hidden;
-
-		& > div {
-			width: 100%;
-		}
-
-		&:hover {
-			box-shadow: none;
-			border: 2px solid black;
-		}
-	}
+.v-item {
+	
+	display: flex;
+	width: 75%;
+	height: 256px;
+	margin: 16px;
+	border-radius: 16px;
+	transition: 0.1s;
+	overflow: hidden;
+	border: 2px solid rgba(0, 0, 0, 0.3);
+	box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+	
 
 	&-image {
-		height: 40%;
+		height: 100%;
+		width: 30%;
 		background-position: center;
 		background-size: cover;
 		background-repeat: no-repeat;
@@ -27,9 +21,8 @@
 
 	&-header {
 		display: flex;
-		flex-direction: column;
-		align-items: start;
-		height: 30%;
+		height: 100%;
+		width: 30%;
 
 		&-name {
 			font-size: 32px;
